Move Scheduler class inside SidekiqScheduler module
Also:
 - Move sidekiq extensions to separate files
