+++ lib/sidekiq-scheduler/extensions/web.rb
@@ -1,0 +1,5 @@
+require 'sidekiq/web' unless defined?(Sidekiq::Web)
+
+Sidekiq::Web.register(SidekiqScheduler::Web)
+Sidekiq::Web.tabs['recurring_jobs'] = 'recurring-jobs'
+Sidekiq::Web.locales << File.expand_path(File.dirname(__FILE__) + '/../../../web/locales')
